backend/requirements.txt
Flask
flask-cors
pdfplumber
python-docx
pytesseract
Pillow
sentence-transformers
scikit-learn
rapidfuzz
opencv-python
numpy
python-dateutil
torch
torchvision
flask-jwt-extended
mysql-connector-python

